+++ application/src/main/java/org/togetherjava/tjbot/commands/Features.java
@@ -125,9 +125,7 @@ public class Features {
         features.add(new WhoIsCommand());
         features.add(new WolframAlphaCommand(config));
         features.add(new AskCommand(config, helpSystemHelper));
-        features.add(new CloseCommand());
-        features.add(new ChangeHelpCategoryCommand(config, helpSystemHelper));
-        features.add(new ChangeHelpTitleCommand(helpSystemHelper));
+        features.add(new HelpThreadCommand(config, helpSystemHelper));
 
         // Mixtures
         features.add(new HelpThreadOverviewUpdater(config, helpSystemHelper));
